requirements.txt

streamlit

flask
langchain
langchain-google-genai
google-generativeai
python-dotenv
langchain_ollama
langchain_community
langchain_groq
groq